lock mechanism for drafts, oop structure for documents, configuration options
git-svn-id: https://weblatex.googlecode.com/svn/trunk@19 7797ea6b-8c6f-a07c-1293-4ddb9af5c572

+++ classes/management/right.class.php
@@ -28,6 +28,7 @@ use weblatex as wl;
 
 require_once( dirname(__DIR__)."/main.class.php" );
 
+    
 
 /** class of representation a right with the database **/
 class right implements \Serializable {
@@ -68,6 +69,12 @@ class right implements \Serializable {
     static function delete( $pnRID, $plForce = false ) {
         if (!is_numeric($pnRID))
             wl\main::phperror( "argument must be a numeric value", E_USER_ERROR );
+        
+        // we check the numeric value of the system rights, so that this right cannot be deleted
+        if (in_array($pnRID, array_values(wl\config::$system_rights), true)) {
+            wl\main::phperror( "system right [".$pnRID."] cannot be deleted", E_USER_NOTICE );
+            throw new \Exception( "system right [".$pnRID."] cannot be deleted" );
+        }
         
         if ($plForce)
             wl\main::getDatabase()->Execute( "DELETE FROM rights WHERE id=?", array($pnRID) );
